# tmp/zrodla/api-docs.yaml
openapi: 3.1.0
info:
  title: RSPO - API
  description: ' '
  version: 1.2.0
servers:
- url: http://api.rspo.gov.pl
  description: Generated server url
security:
- basicAuth: []
...
paths:
...
  /api/placowki/:
    get:
      tags:
      - Placowka
      summary: Retrieves the collection of Placowka resources.
      operationId: getPlacowkaCollection
      parameters:
      - name: wojewodztwo_nazwa
        in: query
        description: full voivodeship name
        required: false
        schema:
          type: string
      - name: wojewodztwo_kod_teryt
        in: query
        description: voivodeship code in TERYT registry
        required: false
        schema:
          type: string
      - name: powiat_nazwa
        in: query
        description: full county name
        required: false
        schema:
          type: string
      - name: powiat_kod_teryt
        in: query
        description: county code in the TERYT registry
        required: false
        schema:
          type: string
      - name: gmina_nazwa
        in: query
        description: full commune name
        required: false
        schema:
          type: string
      - name: gmina_kod_teryt
        in: query
        description: commune code in the TERYT registry
        required: false
        schema:
          type: string
      - name: miejscowosc_nazwa
        in: query
        description: ' full city name'
        required: false
        schema:
          type: string
      - name: miejscowosc_kod_teryt
        in: query
        description: city code in the TERYT registry
        required: false
        schema:
          type: string
      - name: RSPO
        in: query
        description: full RSPO  number
        required: false
        schema:
          type: integer
          format: int32
      - name: NIP
        in: query
        description: full NIP  number
        required: false
        schema:
          type: string
      - name: REGON
        in: query
        description: full REGON  number
        required: false
        schema:
          type: string
      - name: typ_podmiotu_id
        in: query
        description: TypPodmiotu  ID
        required: false
        schema:
          type: integer
          format: int32
      - name: status_publiczno_prawny_id
        in: query
        description: StatusPublicznoPrawny  ID
        required: false
        schema:
          type: integer
          format: int32
      - name: kategoria_ucznia_id
        in: query
        description: KategoriaUcznia  ID
        required: false
        schema:
          type: integer
          format: int32
      - name: etap_edukacji_id
        in: query
        description: EtapEdukacyjny ID
        required: false
        schema:
          type: integer
          format: int32
      - name: zlikwidowana
        in: query
        description: Include liquidated
        required: false
        schema:
          type: boolean
      - name: podmiot_prowadzacy_id
        in: query
        description: 'leading office ID '
        required: false
        schema:
          type: integer
          format: int32
      - name: podmiot_prowadzacy_regon
        in: query
        description: leading office full REGON  number
        required: false
        schema:
          type: string
      - name: page
        in: query
        description: The collection page number
        required: false
        schema:
          type: integer
          format: int32
          minimum: 1
      - name: pageSize
        in: query
        description: The collection page size. Default value is 100
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Placowka collection response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlacowkaDTO"
...
  /api/gminy/:
    get:
      tags:
      - TerytGmina
      summary: Retrieves the collection of TerytGmina resources.
      operationId: getTerytGminaCollection
      parameters:
      - name: powiat.kodTeryt
        in: query
        schema:
          type: array
          items:
            type: string
      - name: nazwa
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
      - name: kodTeryt
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
      - name: page
        in: query
        description: The collection page number
        required: false
        schema:
          type: integer
          format: int32
          minimum: 1
      - name: pageSize
        in: query
        description: The collection page size. Default value is 100
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: TerytGmina collection response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TerytGminaDTO"
...
components:
  schemas:
    ZwiazanieOrganizacyjneDTO:
      type: object
      description: ZwiazanieOrganizacyjne
      properties:
        id:
          type: integer
          format: int32
        nazwa:
          type: string
    TerytWojewodztwoDTO:
      type: object
      description: TerytWojewodztwo
      properties:
        id:
          type: integer
          format: int32
        kodTeryt:
          type: string
          description: "@Groups({\"read\"})"
        nazwa:
          type: string
          description: "* @Groups({\"read\"})"
    TypPodmiotuDTO:
      type: object
      description: TypPodmiotu
      properties:
        id:
          type: integer
          format: int32
        nazwa:
          type: string
    StatusPublicznoPrawnyDTO:
      type: object
      description: StatusPublicznoPrawny
      properties:
        id:
          type: integer
          format: int32
        nazwa:
          type: string
    SpecyfikaSzkolyDTO:
      type: object
      description: SpecyfikaSzkoly
      properties:
        id:
          type: integer
          format: int32
        nazwa:
          type: string
    TerytPowiatDTO:
      type: object
      description: TerytPowiat
      properties:
        id:
          type: integer
          format: int32
        kodTeryt:
          type: string
        nazwa:
          type: string
          description: "@Groups({\"read\"})"
        wojewodztwo:
          $ref: "#/components/schemas/TerytWojewodztwoDTO"
    EtapEdukacyjnyDTO:
      type: object
      description: EtapEdukacyjny
      properties:
        id:
          type: integer
          format: int64
        nazwa:
          type: string
    GeotagDTO:
      type: object
      properties:
        latitude:
          type: string
        longitude:
          type: string
    KategoriaUczniaDTO:
      type: object
      description: KategoriaUcznia
      properties:
        id:
          type: integer
          format: int32
        nazwa:
          type: string
    KsztalcenieDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        nazwa:
          type: string
    KsztalcenieKpssDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        nazwa:
          type: string
    KsztalcenieZawodoweArtystyczneDTO:
      type: object
      properties:
        zawodId:
          type: integer
          format: int32
        zawodNazwa:
          type: string
        specjalnosc:
          type: string
        specjalizacja:
          type: string
    PlacowkaDTO:
      type: object
      description: Placowka
      properties:
        numerRspo:
          type: integer
          format: int32
        dataZalozenia:
          type: string
          format: date-time
        dataRozpoczecia:
          type: string
          format: date-time
        nip:
          type: string
          readOnly: true
        regon:
          type: string
          readOnly: true
        liczbaUczniow:
          type: integer
          format: int32
          readOnly: true
        nazwa:
          type: string
          readOnly: true
        typ:
          $ref: "#/components/schemas/TypPodmiotuDTO"
          readOnly: true
        dyrektorImie:
          type: string
          readOnly: true
        dyrektorNazwisko:
          type: string
          readOnly: true
        statusPublicznoPrawny:
          $ref: "#/components/schemas/StatusPublicznoPrawnyDTO"
        etapyEdukacji:
          type: array
          items:
            $ref: "#/components/schemas/EtapEdukacyjnyDTO"
          readOnly: true
        dataWlaczeniaDoZespolu:
          type: string
          format: date-time
          readOnly: true
        dataLikwidacji:
          type: string
          format: date-time
          readOnly: true
        kategoriaUczniow:
          $ref: "#/components/schemas/KategoriaUczniaDTO"
        specyfikaSzkoly:
          $ref: "#/components/schemas/SpecyfikaSzkolyDTO"
        zwiazanieOrganizacyjne:
          $ref: "#/components/schemas/ZwiazanieOrganizacyjneDTO"
        czyPosiadaObwod:
          type: boolean
          readOnly: true
        czyPosiadaInternat:
          type: boolean
          readOnly: true
        czyDotacjaWPrzyszlymRoku:
          type: boolean
          readOnly: true
        ksztalcenieZawodowe:
          type: array
          items:
            $ref: "#/components/schemas/KsztalcenieDTO"
          readOnly: true
        ksztalcenieZawodoweArtystyczne:
          type: array
          items:
            $ref: "#/components/schemas/KsztalcenieZawodoweArtystyczneDTO"
          readOnly: true
        ksztalcenieWkolegiachPracownikowSluzbSpolecznych:
          type: array
          items:
            $ref: "#/components/schemas/KsztalcenieKpssDTO"
          readOnly: true
        podmiotPrzekazujacyDaneDoRSPO:
          $ref: "#/components/schemas/PodmiotPrzekazujacyDaneDTO"
          readOnly: true
        podmiotProwadzacy:
          type: array
          items:
            $ref: "#/components/schemas/PodmiotProwadzacyDTO"
          readOnly: true
        podmiotNadrzedny:
          type: string
        wojewodztwo:
          type: string
          readOnly: true
        wojewodztwoKodTERYT:
          type: string
          readOnly: true
        powiat:
          type: string
          readOnly: true
        powiatKodTERYT:
          type: string
          readOnly: true
        gmina:
          type: string
          readOnly: true
        gminaKodTERYT:
          type: string
          readOnly: true
        gminaRodzaj:
          type: string
          readOnly: true
        miejscowosc:
          type: string
          readOnly: true
        miejscowoscKodTERYT:
          type: string
          readOnly: true
        ulica:
          type: string
          readOnly: true
        ulicaKodTERYT:
          type: string
          readOnly: true
        numerBudynku:
          type: string
          readOnly: true
        numerLokalu:
          type: string
          readOnly: true
        kodPocztowy:
          type: string
          readOnly: true
        poczta:
          type: string
          readOnly: true
        geolokalizacja:
          type: array
          items:
            $ref: "#/components/schemas/GeotagDTO"
          readOnly: true
        telefon:
          type: string
          readOnly: true
        email:
          type: string
          readOnly: true
        stronaInternetowa:
          type: string
          readOnly: true
        adresDoKorespondecjiMiejscowosc:
          type: string
          readOnly: true
        adresDoKorespondecjiUlica:
          type: string
          readOnly: true
        adresDoKorespondecjiNumerBudynku:
          type: string
          readOnly: true
        adresDoKorespondecjiNumerLokalu:
          type: string
          readOnly: true
        adresDoKorespondecjiKodPocztowy:
          type: string
          readOnly: true
        adresDoKorespondecjiPoczta:
          type: string
          readOnly: true
        placowkiPodrzedne:
          type: string
          readOnly: true
        liczbaOddzialow:
          type: integer
          format: int32
          readOnly: true
        rodzajSzkolyPlacowki:
          $ref: "#/components/schemas/RodzajSzkolyPlacowkiDTO"
          readOnly: true
    PodmiotProwadzacyDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        nazwa:
          type: string
        regon:
          type: string
        typ:
          $ref: "#/components/schemas/TypPodmiotuDTO"
    PodmiotPrzekazujacyDaneDTO:
      type: object
      properties:
        nazwa:
          type: string
        regon:
          type: string
        typ:
          $ref: "#/components/schemas/TypPodmiotuDTO"
    RodzajSzkolyPlacowkiDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        nazwa:
          type: string
    TerytGminaDTO:
      type: object
      description: TerytGmina
      properties:
        id:
          type: integer
          format: int32
        kodTeryt:
          type: string
        nazwa:
          type: string
        active:
          type: boolean
        powiat:
          $ref: "#/components/schemas/TerytPowiatDTO"
        rodzajGminy:
          $ref: "#/components/schemas/TerytRodzajGminyDTO"
...
    TerytRodzajGminyDTO:
      type: object
      description: TerytRodzajGminy
      properties:
        nazwa:
          type: string
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic